+++ netskope_cls/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.2.0
+## Added
+- Added support for Client status events, To pull and ingest this event type update your Netskope Tenant version to 1.4.0.
+
+# 2.1.0
+## Added
+- Added support for device and content alerts. To pull and ingest these alert types update your Netskope Tenant version to 1.3.0.
+
 # 2.0.0
 ## Changed
 - The Netskope CLS plugin has been restructured and is now available in the Default repository.