+++ netskope_provider/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.4.0
+## Added
+- Added support for Client status events. To pull and ingest this event type update your CLS Netskope plugin version to 2.2.0.
+
+# 1.3.0
+## Changed
+- Enhanced field learning workflow with updated plugin to identify fields and their data types.
+
+# 1.2.0
+## Added
+- Added support for device and content alerts.
+
 # 1.1.0
 ## Changed
 - Enhanced pulling workflow.